rust-sdl2 update

+++ examples/triangle.rs
@@ -132,7 +132,7 @@ fn make_window() -> sdl2::video::Window {
   sdl2::video::gl_set_attribute(sdl2::video::GLAttr::GLContextMinorVersion, 0);
   sdl2::video::gl_set_attribute(
     sdl2::video::GLAttr::GLContextProfileMask,
-    sdl2::video::GLProfile::GLCoreProfile as isize
+    sdl2::video::GLProfile::GLCoreProfile as i32,
   );
 
   sdl2::video::Window::new(
